int to long for countByStudyIdIn

+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/repository/VariantSourceEntityRepository.java
@@ -39,6 +39,6 @@ public interface VariantSourceEntityRepository extends MongoRepository<VariantSo
 
     List<VariantSourceEntity> findByStudyIdIn(List<String> studyIds, Pageable pageable);
 
-    int countByStudyIdIn(List<String> studyIds);
+    long countByStudyIdIn(List<String> studyIds);
 
 }
